requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.105.0
uvicorn>=0.25.0
requests>=2.31.0
mysql-connector-python>=9.0.0
python-multipart>=0.0.9
streamlit>=1.28.0
passlib>=1.7.4
PyMySQL>=1.1.0